pyproject.toml
[tool.poetry]
name = "poetry"
version = "1.3.2"
description = "automated-shorts"
authors = [
    "Keith Carringotn <keithlewiscarrington@gmail.com>",
]

# Requirements
[tool.poetry.dependencies]
python = "^3.8"
poetry-core = "^1.1.0"
poetry-plugin-export = "^1.0.6"
httplib2 = "^0.20.4"
nltk = "^3.7"
numpy = "^1.23.2"
oauth2client = "^4.1.3"
notebook = "^6.4.12"
google-api-python-client = "^2.58.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
psycopg2 = "^2.9.3"
requests = "^2.28.1"
pycurl = "^7.45.1"
pandas = "^1.4.4"
boto3 = "^1.24.68"
mysql-connector = "^2.2.9"
sentence-transformers = "^2.2.2"
Markdown = "^3.4.1"
md-mermaid = "^0.1.1"
prefect = "^2.7.12"
pulumi-aws = "^5.29.1"
pytrends = "^4.9.0"
googlenews = "^1.6.6"
newspaper3k = "^0.2.8"
openai = "^0.27.0"
wand = "^0.6.11"
moviepy = "^1.0.3"
tiktokapi-pyppeteer = "^3.7.9"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py37']
preview = true
force-exclude = '''
.*/setup\.py$
'''
